[Costing] Use the handle pattern
