+++ plutus-core/plutus-core.cabal
@@ -138,40 +138,24 @@ library
         prelude
         common
     other-modules:
-        Language.PlutusCore.Pretty.ConfigName
-        Language.PlutusCore.Core.Type
-        Language.PlutusCore.Core.Plated
+        Language.PlutusCore.Analysis.Definitions
+        Language.PlutusCore.Constant.Apply
+        Language.PlutusCore.Constant.Function
+        Language.PlutusCore.Constant.Meaning
+        Language.PlutusCore.Constant.Typed
+        Language.PlutusCore.Core.Instance
         Language.PlutusCore.Core.Instance.Eq
+        Language.PlutusCore.Core.Instance.Pretty
         Language.PlutusCore.Core.Instance.Pretty.Classic
         Language.PlutusCore.Core.Instance.Pretty.Common
         Language.PlutusCore.Core.Instance.Pretty.Default
         Language.PlutusCore.Core.Instance.Pretty.Plc
         Language.PlutusCore.Core.Instance.Pretty.Readable
-        Language.PlutusCore.Core.Instance.Pretty
         Language.PlutusCore.Core.Instance.Recursive
-        Language.PlutusCore.Core.Instance
-        Language.PlutusCore.Constant.Apply
-        Language.PlutusCore.Constant.Meaning
-        Language.PlutusCore.Constant.Function
-        Language.PlutusCore.Constant.Typed
+        Language.PlutusCore.Core.Plated
+        Language.PlutusCore.Core.Type
         Language.PlutusCore.DeBruijn.Internal
-        Language.PlutusCore.Lexer.Type
         Language.PlutusCore.Eq
-        Language.PlutusCore.Mark
-        Language.PlutusCore.Parser.Internal
-        Language.PlutusCore.Pretty.Classic
-        Language.PlutusCore.Pretty.ConfigName
-        Language.PlutusCore.Pretty.Default
-        Language.PlutusCore.Pretty.Plc
-        Language.PlutusCore.Pretty.PrettyConst
-        Language.PlutusCore.Pretty.Readable
-        Language.PlutusCore.Pretty.Utils
-        Language.PlutusCore.Universe.Core
-        Language.PlutusCore.Universe.Default
-        Language.PlutusCore.Size
-        Language.PlutusCore.TypeCheck.Internal
-        Language.PlutusCore.TypeCheck
-        Language.PlutusCore.Analysis.Definitions
         Language.PlutusCore.Examples.Data.InterList
         Language.PlutusCore.Examples.Data.Shad
         Language.PlutusCore.Examples.Data.TreeForest
@@ -182,35 +166,52 @@ library
         Language.PlutusCore.Generators.Internal.TypeEvalCheck
         Language.PlutusCore.Generators.Internal.TypedBuiltinGen
         Language.PlutusCore.Generators.Internal.Utils
+        Language.PlutusCore.Lexer.Type
+        Language.PlutusCore.Mark
         Language.PlutusCore.Parsable
+        Language.PlutusCore.Parser.Internal
+        Language.PlutusCore.Pretty.Classic
+        Language.PlutusCore.Pretty.ConfigName
+        Language.PlutusCore.Pretty.ConfigName
+        Language.PlutusCore.Pretty.Default
+        Language.PlutusCore.Pretty.Plc
+        Language.PlutusCore.Pretty.PrettyConst
+        Language.PlutusCore.Pretty.Readable
+        Language.PlutusCore.Pretty.Utils
+        Language.PlutusCore.Size
+        Language.PlutusCore.TypeCheck
+        Language.PlutusCore.TypeCheck.Internal
+        Language.PlutusCore.Universe.Core
+        Language.PlutusCore.Universe.Default
 
         Language.PlutusIR.Analysis.Dependencies
         Language.PlutusIR.Analysis.Usages
-        Language.PlutusIR.Compiler.Let
         Language.PlutusIR.Compiler.Datatype
+        Language.PlutusIR.Compiler.Error
+        Language.PlutusIR.Compiler.Let
+        Language.PlutusIR.Compiler.Lower
         Language.PlutusIR.Compiler.Provenance
         Language.PlutusIR.Compiler.Recursion
         Language.PlutusIR.Compiler.Types
-        Language.PlutusIR.Compiler.Lower
-        Language.PlutusIR.Compiler.Error
         Language.PlutusIR.Normalize
         Language.PlutusIR.TypeCheck.Internal
 
+        Language.UntypedPlutusCore.Analysis.Definitions
+        Language.UntypedPlutusCore.Check.Uniques
         Language.UntypedPlutusCore.Core
         Language.UntypedPlutusCore.Core.Instance
+        Language.UntypedPlutusCore.Core.Instance.CBOR
         Language.UntypedPlutusCore.Core.Instance.Eq
+        Language.UntypedPlutusCore.Core.Instance.Flat
         Language.UntypedPlutusCore.Core.Instance.Pretty
         Language.UntypedPlutusCore.Core.Instance.Pretty.Classic
         Language.UntypedPlutusCore.Core.Instance.Pretty.Default
         Language.UntypedPlutusCore.Core.Instance.Pretty.Plc
         Language.UntypedPlutusCore.Core.Instance.Pretty.Readable
         Language.UntypedPlutusCore.Core.Instance.Recursive
-        Language.UntypedPlutusCore.Core.Instance.CBOR
-        Language.UntypedPlutusCore.Core.Instance.Flat
-        Language.UntypedPlutusCore.Core.Type
         Language.UntypedPlutusCore.Core.Plated
-        Language.UntypedPlutusCore.Analysis.Definitions
-        Language.UntypedPlutusCore.Check.Uniques
+        Language.UntypedPlutusCore.Core.Type
+        Language.UntypedPlutusCore.Evaluation.Machine.Cek.Internal
         Language.UntypedPlutusCore.Mark
         Language.UntypedPlutusCore.Rename.Internal
         Language.UntypedPlutusCore.Size
